refactor #164: CustomerRefund & StoreRefund 통합
