+++ src/main/java/com/irum/come2us/domain/refund/application/service/RefundService.java
@@ -5,10 +5,13 @@ import com.irum.come2us.domain.order.domain.entity.OrderDetail;
 import com.irum.come2us.domain.order.domain.repository.OrderDetailRepository;
 import com.irum.come2us.domain.order.domain.repository.OrderRepository;
 import com.irum.come2us.domain.refund.domain.entity.Refund;
+import com.irum.come2us.domain.refund.domain.entity.enums.RefundStatus;
 import com.irum.come2us.domain.refund.domain.repository.RefundRepository;
-import com.irum.come2us.domain.refund.presentation.dto.request.CustomerRefundCreateRequest;
-import com.irum.come2us.domain.refund.presentation.dto.response.CustomerRefundDetailResponse;
+import com.irum.come2us.domain.refund.presentation.dto.request.RefundCreateRequest;
+import com.irum.come2us.domain.refund.presentation.dto.request.StoreRefundStatusRequest;
+import com.irum.come2us.domain.refund.presentation.dto.response.*;
 import com.irum.come2us.global.presentation.advice.exception.CommonException;
+import com.irum.come2us.global.presentation.advice.exception.errorcode.OrderErrorCode;
 import com.irum.come2us.global.presentation.advice.exception.errorcode.RefundErrorCode;
 import com.irum.come2us.global.util.MemberUtil;
 import java.util.List;
@@ -22,26 +25,67 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 @Transactional
 @Slf4j
-public class CustomerRefundService {
+public class RefundService {
+
     private final MemberUtil memberUtil;
     private final RefundRepository refundRepository;
     private final OrderRepository orderRepository;
     private final OrderDetailRepository orderDetailRepository;
 
-    public void createRefund(UUID orderId, CustomerRefundCreateRequest request) {
+    // Customer
+    public void createRefund(UUID orderId, RefundCreateRequest request) {
         assertNoRefundExistsByOrder(orderId);
         Order order = getValidOrder(orderId);
         refundRepository.save(
                 Refund.create(
-                        request.reason(), request.description(), order.getPayment().getAmount()));
+                        request.reason(),
+                        request.description(),
+                        order.getPayment().getAmount())); // Payment 도메인 Getter 삽입 필요
     }
 
     @Transactional(readOnly = true)
-    public CustomerRefundDetailResponse findRefundDetail(UUID orderId) {
+    public RefundDetailResponse findRefundDetail(UUID orderId) {
         Order order = getValidOrderWithAddress(orderId);
         Refund refund = getValidRefund(orderId);
         List<OrderDetail> orderDetails = orderDetailRepository.findAllByOrder(order);
-        return CustomerRefundDetailResponse.of(order, orderDetails, refund);
+        return RefundDetailResponse.of(order, orderDetails, refund);
+    }
+
+    // Owner
+    @Transactional(readOnly = true)
+    public StoreRefundListResponse findRefundListByStatus(RefundStatus status) {
+        List<Refund> refunds = refundRepository.findByRefundStatus(status);
+
+        var orderList =
+                refunds.stream()
+                        .map(
+                                refund -> {
+                                    var order = refund.getOrder();
+                                    var products = List.<RefundProductList>of();
+
+                                    return new RefundOrderList(
+                                            order.getOrderId(),
+                                            order.getDeliveryAddress().getRecipientName(),
+                                            order.getCreatedAt(),
+                                            refund.getCreatedAt(),
+                                            refund.getRefundStatus(),
+                                            refund.getPrice(),
+                                            products);
+                                })
+                        .toList();
+
+        int total = refunds.stream().mapToInt(Refund::getPrice).sum();
+
+        return new StoreRefundListResponse(orderList, String.valueOf(total), false, null);
+    }
+
+    public void changeRefundStatus(UUID refundId, StoreRefundStatusRequest request) {
+        Refund refund =
+                refundRepository
+                        .findById(refundId)
+                        .orElseThrow(() -> new CommonException(RefundErrorCode.REFUND_NOT_FOUND));
+
+        refund.updateStatus(request.refundStatus());
     }
 
     private void assertNoRefundExistsByOrder(UUID orderId) {
