--- src/main/java/com/irum/come2us/domain/refund/presentation/dto/request/StoreRefundCreateRequest.java
@@ -1,8 +1,0 @@
-package com.irum.come2us.domain.refund.presentation.dto.request;
-
-import com.irum.come2us.domain.refund.domain.entity.enums.RefundReason;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-
-public record StoreRefundCreateRequest(
-        @NotNull RefundReason refundReason, @NotBlank String description) {}
